update feature upload large file

+++ src/main/java/com/example/demo/service/ProfileService.java
@@ -1,9 +1,8 @@
 package com.example.demo.service;
-
 import com.example.demo.Response.CommonResponse;
 import com.example.demo.exception.BaseException;
 import com.example.demo.exception.ProfileException;
-import com.example.demo.model.ImageData;
+import com.example.demo.model.Files;
 import com.example.demo.model.Profile;
 import com.example.demo.model.User;
 import com.example.demo.repository.ProfileRepository;
@@ -12,7 +11,6 @@ import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
-
 import java.io.IOException;
 import java.util.List;
 import java.util.Optional;
@@ -23,9 +21,9 @@ public class ProfileService {
     private final ProfileRepository profileRepository;
     private final StorageService service;
     private  final UserRepository userRepository;
-    private  final  Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
 
     private Integer getUserId(){
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
         User user =  (User) authentication.getPrincipal();
         return user.getId();
     }
@@ -39,12 +37,12 @@ public class ProfileService {
     public CommonResponse createProfile(MultipartFile file, String fullname,String detail) throws IOException {
         User userAuth = authService.getInfo();
         Optional<User> user = userRepository.findById(userAuth.getId());
-        ImageData imageData =  service.uploadImageToFileSystem(file);
+        Files files =  service.uploadImageToFileSystem(file);
         Profile profile = new Profile();
         profile.setFullname(fullname);
         profile.setDetail(detail);
         user.ifPresent(profile::setUser);
-        profile.setImageData(imageData);
+        profile.setFiles(files);
         profileRepository.save(profile);
         return new CommonResponse("Create profile successfully");
     }
@@ -54,13 +52,14 @@ public class ProfileService {
         profile.setFullname(fullname);
         profile.setDetail(detail);
         if (file != null){
-            service.removeFile(profile.getImageData().getId());
-            ImageData imageData =  service.uploadImageToFileSystem(file);
-            profile.setImageData(imageData);
+            service.removeFile(profile.getFiles().getId());
+            Files files =  service.uploadImageToFileSystem(file);
+            profile.setFiles(files);
         }
         profileRepository.save(profile);
         return profile;
     }
+
     public CommonResponse removeProfile(Integer profileId) throws BaseException {
         if(profileId == null){
             throw  ProfileException.ProfileIdIsEmpty();
@@ -70,9 +69,12 @@ public class ProfileService {
             throw  ProfileException.ProfileNotFound();
         }
         Profile profile = profileOpt.get();
-        service.removeFile(profile.getImageData().getId());
-        profileRepository.delete(profile);
-        return new CommonResponse("Delete Profile Success");
+        if(service.removeFile(profile.getFiles().getId())){
+            profileRepository.delete(profile);
+            return new CommonResponse("Delete Profile Success");
+        }
+        return  new CommonResponse("Delete Profile Fail");
+
     }
 
     public List<Profile> getProfile(){
